% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubigeo_helpers.R
\name{ubigeo_departments}
\alias{ubigeo_departments}
\title{Tabla de referencia de departamentos / Department reference table}
\usage{
ubigeo_departments()
}
\value{
data.table con codigos y nombres de departamentos / data.table with department codes and names
}
\description{
Tabla de referencia de departamentos / Department reference table
}
\examples{
deps <- ubigeo_departments()
}
