% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubigeo_helpers.R
\name{ubigeo_get_province}
\alias{ubigeo_get_province}
\title{Extraer provincia de UBIGEO / Extract province from UBIGEO}
\usage{
ubigeo_get_province(ubigeo)
}
\arguments{
\item{ubigeo}{Vector de códigos UBIGEO de 6 dígitos / Vector of 6-digit UBIGEO codes}
}
\value{
Vector con códigos de provincia (4 dígitos: DDPP) / Vector with province codes (4 digits: DDPP)
}
\description{
Extraer provincia de UBIGEO / Extract province from UBIGEO
}
\examples{
ubigeo_get_province("150101")  # "1501" (Lima province)
}
