% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubigeo_helpers.R
\name{ubigeo_pad}
\alias{ubigeo_pad}
\title{Agregar ceros a la izquierda en UBIGEO / Pad UBIGEO with leading zeros}
\usage{
ubigeo_pad(ubigeo)
}
\arguments{
\item{ubigeo}{Vector de codigos UBIGEO que pueden estar sin ceros / Vector of UBIGEO codes that may lack zeros}
}
\value{
Vector con codigos UBIGEO de 6 digitos / Vector with 6-digit UBIGEO codes
}
\description{
Agregar ceros a la izquierda en UBIGEO / Pad UBIGEO with leading zeros
}
\examples{
ubigeo_pad("10101")   # "010101"
ubigeo_pad("150101")  # "150101"
}
