% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubigeo_helpers.R
\name{vigilaPeru-imports}
\alias{vigilaPeru-imports}
\title{Funciones auxiliares para manejo de codigos UBIGEO / Helper functions for UBIGEO codes}
\description{
Funciones para trabajar con codigos UBIGEO (Ubicacion Geografica) del Peru.
Los codigos UBIGEO tienen formato DDPPTT donde:
\itemize{
\item DD: Departamento (2 digitos)
\item PP: Provincia (2 digitos)
\item TT: Distrito (2 digitos)
}

Functions to work with Peru's UBIGEO (Geographic Location) codes.
UBIGEO codes have format DDPPTT where:
\itemize{
\item DD: Department (2 digits)
\item PP: Province (2 digits)
\item TT: District (2 digits)
}
}
\keyword{internal}
