% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{vp_cache_info}
\alias{vp_cache_info}
\title{Informacion del cache / Cache information}
\usage{
vp_cache_info()
}
\value{
data.frame con informacion del cache / data.frame with cache information
}
\description{
Muestra informacion sobre el uso del cache de vigilaPeru.

Shows information about vigilaPeru cache usage.
}
\examples{
vp_cache_info()
}
