% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubigeo_helpers.R
\name{ubigeo_is_valid}
\alias{ubigeo_is_valid}
\title{Validar formato de UBIGEO / Validate UBIGEO format}
\usage{
ubigeo_is_valid(ubigeo)
}
\arguments{
\item{ubigeo}{Vector de codigos UBIGEO / Vector of UBIGEO codes}
}
\value{
Vector logico indicando si cada codigo es valido / Logical vector indicating if each code is valid
}
\description{
Validar formato de UBIGEO / Validate UBIGEO format
}
\examples{
ubigeo_is_valid(c("150101", "15", "999999", "ABC"))
}
