% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{vp_cache_info}
\alias{vp_cache_info}
\title{Información del caché / Cache information}
\usage{
vp_cache_info()
}
\value{
data.frame con información del caché / data.frame with cache information
}
\description{
Muestra información sobre el uso del caché de vigilaPeru.

Shows information about vigilaPeru cache usage.
}
\examples{
vp_cache_info()
}
